requirements-dev.txt
-r requirements.txt
flake8==3.5.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
atomicwrites==1.1.5
attrs==18.1.0
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.4
pytest==3.6.3
six==1.11.0
pytest-django==3.3.2
coverage==4.5.1
selenium==3.13.0
